Add and configure User policy

+++ app/controllers/profiles/accounts_controller.rb
@@ -1,6 +1,11 @@
 # frozen_string_literal: true
 
 class Profiles::AccountsController < ApplicationController
+  include AuthHelper
+
+  before_action :authenticate_user!
+  before_action :set_user, except: %i[unlink]
+
   def show; end
 
   def unlink
@@ -9,8 +14,15 @@ class Profiles::AccountsController < ApplicationController
 
     return render_404 unless identity
 
-    identity.destroy
+    identity.destroy if unlink_provider_allowed?(provider)
 
     redirect_to profile_account_path
+  end
+
+  private
+
+  def set_user
+    @user = current_user
+    authorize @user
   end
 end
